refactor: split translate.js into components and utils
